Update expose command for kubectl cli
At least in version v1.11.1, the format for exposing a Kubernetes application is: 

Usage:
  kubectl expose (-f FILENAME | TYPE NAME) [--port=port] [--protocol=TCP|UDP] [--target-port=number-or-name]
[--name=name] [--external-ip=external-ip-of-service] [--type=type] [options]

The "--port=3000" flag should come before "--type=NodePort" otherwise it isn't recognised. Similarly, the "--name" flag names the service which will be the same as the deployment's name by default anyway. Having it there for me for some reason threw an error: 

Error from server (NotFound): deployments.extensions "—-name=deploy-react-kubernetes-service" not found.

I just removed it and it worked.